+++ pact_testgen/pact_testgen.py
@@ -81,19 +81,22 @@ def convert_to_test_cases(pact: Pact, base_class: str) -> TestFile:
         # We need a hashable collection to key on, but also need to rememeber
         # name order so that test case names are deterministic based on the
         # order defined in the Pact contract.
-        provider_state_key = frozenset(
-            [ps.full_name() for ps in interaction.providerStates]
-        )
+        if interaction.providerStates is not None:
+            provider_state_key = frozenset(
+                [ps.full_name() for ps in interaction.providerStates]
+            )
+        else:
+            provider_state_key = None
         provider_states_interactions[provider_state_key].append(interaction)
 
     cases = []
 
     for (
-        _,
+        provider_state_key,
         interactions,
     ) in provider_states_interactions.items():
-        cases.append(
-            TestCase(
+        if provider_state_key is not None:
+            case = TestCase(
                 # Interactions have been grouped by the set of
                 # provider states, and we know there is at least
                 # one interaction, so we can use the provider states
@@ -103,7 +106,12 @@ def convert_to_test_cases(pact: Pact, base_class: str) -> TestFile:
                 ],
                 test_methods=[_build_method_args(i) for i in interactions],
             )
-        )
+        else:
+            case = TestCase(
+                provider_state_names=[],
+                test_methods=[_build_method_args(i) for i in interactions],
+            )
+        cases.append(case)
 
     return TestFile(
         base_class=base_class,