+++ HISTORY.rst
@@ -11,6 +11,7 @@ History
 * Adds option to merge changes to provider state file (Python 3.9 only).
 * Include structured provider state parameters from Pact v3 in provider
   state function names.
+* Support null provider state, i.e. pact.given(None).
 
 0.2.1 (2021-09-01)
 ------------------